+++ src/fineman/heavy_light_partition.py
@@ -1,7 +1,7 @@
 from typing import Tuple,Set
 import random as rand
 from math import ceil, log
-from fineman.helper_functions import compute_reach
+from src.fineman.helper_functions import compute_reach
 
 def heavy_light_partition(graph, neg_edges, negative_subset, rho, c: int, seed=None) -> Tuple[Set[int], Set[int]]:
     if seed is not None: rand.seed(seed)
