+++ src/fineman/independent_set_or_crust.py
@@ -1,6 +1,6 @@
-from fineman.helper_functions import *
-from fineman.rand_is import *
-from fineman.heavy_light_partition import *
+from src.fineman.helper_functions import *
+from src.fineman.rand_is import *
+from src.fineman.heavy_light_partition import *
 import time
 from math import ceil, log2
 import random as rand
