+++ src/tests/test_elimination_by_hop_reduction.py
@@ -1,7 +1,7 @@
 import pytest
 
-from fineman.elimination_by_hop_reduction import construct_h, elimination_of_r_remote_edges_by_hop_reduction
-from utils import load_test_case, NegativeCycleError
+from src.fineman.elimination_by_hop_reduction import construct_h, elimination_of_r_remote_edges_by_hop_reduction
+from src.utils import load_test_case, NegativeCycleError
 
 TESTDATA_FILEPATH = "src/tests/test_data/graphs/"
