+++ src/tests/test_is_or_crust.py
@@ -1,6 +1,6 @@
 import pytest
-from fineman import find_is_or_crust
-from utils.load_test_case import load_test_case
+from src.fineman import find_is_or_crust
+from src.utils.load_test_case import load_test_case
 
 TESTDATA_FILEPATH = "src/tests/test_data/graphs/"
