+++ src/tests/test_preprocessing.py
@@ -1,6 +1,6 @@
 import pytest
-from fineman.preprocessing import *
-from utils.load_test_case import load_test_case
+from src.fineman.preprocessing import *
+from src.utils.load_test_case import load_test_case
 
 TESTDATA_FILEPATH = "src/tests/test_data/graphs/"
